backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.30
chromadb>=0.4.0
openai>=1.10.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
tiktoken>=0.5.0
pandas>=2.0.0
numpy>=1.24.0